Fixing some remaining bugs in interface exchange and parallel ghosting.  ALL parallel tests done for "make check" now work, except for some of the iMeshP functions.  I still see some errors with some other files, will track those down next.
parallel/pcomm_unit: added ability to serially test for more general
  cases (allow input of distrib flag, instead of hardcoding it); also
  now points to source directory if using out-of-source build

parallel/MBParallelComm: 
- reserve MAX_SHARING_PROCS spaces in buffer and proc vectors; this
works around a bug, where, when the vector was resized due to adding
another processor as a result of ghost communication, some pointers
into the old vector became stale; this results in a maximum of
2*MAX_SHARING_PROCS*sizeof(std::vector<MBEntityHandle>) wasted space
on a processor, which is ok for now.
- make init function to call from both constructors
- got rid of unused myBuffers vector
- added a few asserts to make sure we weren't trying to unpack -1 ints
- adjusted list_entities argument processing for consistency with
MBCore behavior
- changed get_sharing_data and made another variant of same, that is
able to perform booleans on results, e.g. give me the processors
sharing all these entities; substituted some repeated code doing this
type of evaluation with call to this function
- added tag_iface_entities function, to correctly tag the non-vertex
entities in interface sets according to the pstatus of the interface
set; this is done after exchanging ghost cells call returns, rather
than during the function itself, because some entities previously
though to be interface ents are removed during that process
- fixed a few memory leaks and uninitialized variables
- for interface sets, assign at least part of the shared handles tag
to reflect local handle of set
- broke up consistency checking tests, and saved strings indicating
reason for failed checks, for individual entities, and repeats these
reasons in output when consistency check fails
- when DEBUG_COMM is defined, use barriers at the end of major rounds
of communication during ghost exchange

test/64bricks_512hex.h5m: changing, in case this was generated while 
HDF5 wasn't working (probably not, but just to be safe).



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3056 6162379e-bd28-0410-9a7d-b7f4dcfcad3c