Add support for reading mesh generated for use in ABAQUS. [Note: these files typiclly use the extension ".inp"        but that seemed to generic so I set up automatic        support for ".abq" instead.]
This mesh format implements a part/assembly/instance
paradigm with nodes & elements collected in node sets
and element sets.  This was implemented in MOAB using
tags and meshsets (but no parent-child) relationships.
For example, assembly meshsets contain instance meshsets
and instance meshsets are tagged with the handle of the
assembly meshset that contains them.  At the moment an 
extensive set of cross-referencing was done with this
general technique.  This should probably be documented
in more detail for this implementation, but could probably
have portions of it abstracted in order to support other
mesh formats with a similar paradigm.

At this point, the reader supports a limited set of
keywords, largely for determining which material is
in each mesh element.

The reader also only supports a limited set of 
ABAQUS element types, but extending for additional types
is straightforward.  The corollary is that the purpose
of this reader is so limited that the type of element
rarely matters beyond mapping it onto the correct 
MBEntityType and knowning the correct connectivity size.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3114 6162379e-bd28-0410-9a7d-b7f4dcfcad3c