Use faceting tolerance as the tolerance to determine whether a boundary case should be tested during point membership.  If the current point is within faceting tolerance of a boundary, assume that it is a boundary case.
Need to ensure that faceting tolerance is known in more circumstances.  When
reading a CGM geometry a faceting tolerance can be provided and passed to the
load_file routine as an option.  This will tag the resultant MOAB mesh with 
the faceting tolerance.  We now read the faceting tolerance from the mesh and
use that as the assumed value.

Now also allow the calling application to query the faceting tolerance along
with other settings.




git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3965 6162379e-bd28-0410-9a7d-b7f4dcfcad3c