add utilities to create the visu mesh for a spectral element, also to project nodes on a sphere rename IntxUtils as CslamUtils add a dependency to mbcoupler because the spectral element is defined there. it may move to a more central place.
add a test for this visu mesh

this visu mesh might evolve into a spectral element in moab (see yesterday
discusion). 

also, the way it was created is not the most efficient, but it is the easiest 
to implement. 
basically, compute the gl points, create nodes (for each gl point), mesh into
quads, and then MERGE :)

It would be probably more efficient to create "spectral" nodes on each edge 
first, then look at the orientation of each edge in each quad, create the 
interior nodes, and then mesh interior of each coarse quad.

It would avoid the merging step.

 


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5849 6162379e-bd28-0410-9a7d-b7f4dcfcad3c