default value for PARALLEL_PARTITION tag
both visit and mbpart like to have a default value for this tag.
Otherwise, visit would not be able to "see" parallel partition in an h5m file.
mbpart would also fail to repartition. mbpart would have been easy to fix, but
visit is a different story. So, it is better to have a default value.
This tag gets created right now either by mbpart, or by "read_nc"-like reads in
parallel (CAM, FV or MPAS reads on more processors)