add a crystal router example
crystal router is one of the building blocks of parallel
infrastructure in MOAB, so in my opinion, it deserves
an example.
In this example, it is shown how to gather-scatter homogeneous
data between processors, with one call. It is assumed that the
communication matrix is relatively sparse, and this is the case
for most applications.