changes for 64bit handle on 32 bit machine
so we can configure on 32 bit machines with 64 bit handles
in that case, the EntityHandle is usually converted to unsigned long long
(if the compiler/system allows it), and it has size 8 bytes

Not very likely somebody needs to do this
really, why would we need a lot of handles on a 32 bit machine?
Maybe on intrepid, but intrepid is retired, way too old.

Still, these are for consistency