modify the large mesh example to use an opaque tag for resolving sharing
so, instead of using the global id tag to resolve sharing, the
example will use a new tag, opaque, sizeof(long), because it will be
used directly like that for the crystal router
(the tag will be created as opaque, but it will simply convert a long
to bytes )

We are using this to generate a file with more than 1 B vertices

Before, we were using an integer tag (global id), which had its
limitations.

The same tag type is used now when reading a hdf5 file in parallel, to
create the file_id_tag. This is used (of course) to resolve shared
vertices.

The LoadPartial example is modified to show the how to read that special
tag created, and interpret it as long.
(We are doing actually a casting when we read the tag, we know its
type, but when we read we just pass the pointer to beginning
of the long array: &valsTag[0])

under the hood, a conversion from bytes to long takes place
(the reverse conversion takes place when we actually set the tag, from long
to opaque, bytes)