Make order of availableList deterministic.
If sequence of creation and deletion of entities is the same for each process run, entities handlers will have the same value independent on operations system and program execution. Note that set of pointers would not iterate over in a deterministic manner unless classes unique ids are compared not pointers itself.