Remove usage of Hypre matrices from our weight matrix computation algorithm.
Considering the limited use case for a full parallel matrix, we will instead
use just the local copy of the matrix in each process, with added knowledge
of shared row contributions that will be used in reduction after matvec
computations.

And for that purpose, we instead use Eigen3::SparseMatrix datastructures.