The errors in serial and parallel occur due to variation in number of overlap elements that are available on the local partition.
To better investigate and to optimize the local least squares
solve, we have now re-implemented the ForceConsistencyConservation3
routine in MOAB and use Eigen3 matrices to perform the solves.
The assembly can still be optimized further.

For now, on the default mesh combinations, the remapper writes
out the dC/dCCt matrices for elem=504, which has large errors
since the source element extends between partitions.
Once we have a solution, we should be able to verify that
the matrices match exactly.