reduce the default size for sequence
the  problem is relevant for relatively small meshes, when we have many
dense tags (for example climate applications)
We can quickly run out of memory :(
A better fix would be to create in bulk the cells in coverage mesh, not in a loop,
one by one.
Then the default sequence size would not matter, as we can create
a sequence of desired size, which should be small. We should look also at migrate,
if we are creating sequences of good size.
So if we have then 37 dense tags of size 16 would not run out of memory
on gnep (that has only 4 Gb of memory)