various fixes for parallel infrastructure
MPI_Reduce for MPI_SUM operation should have the global accumulated
variable initialized to 0. It is failing otherwise on large tasks count
with intel compiler and openmpi implementation

It is a safe thing to do anyway
It was needed also when we computed the L2 norm for differences
between MCT and MOAB driver fields

Another fix is for extra tags that need to be exchanged when we
ghost source or target sets for remapping

GLOBAL_ID is a tag that is always needed in all our workflows.
For spectral elements, when we define the GLOBAL_DOFS tag on quads,
that will need to be exchanged for ghost layers too, otherwise a lot of the
logic for spectral cells does not work.
As a reminder, GLOBAL_DOFS tag holds the local degrees of freedom for each
spectral element, in lexicographic ordering. For np=4, it is a size 16 integer tag

It is an integer, so it cannot be larger than 2**31 - 1, due to integer
overflow
2**31-1 ~= 2B. This is a hard limit on HOMME too, it may need to be changed
in the future.